Add Nullness annotations (JSR 305)
Summary:
Now all parameters have the corresponding @Nonnull and @Nullable
annotations.

If the parameter was used assuming it was not null, a @Nonnull
annotation was added.

If the parameter might be null without breaking everything (example:
object.Equals or dictionary lookup), it was tagged with @Nullable.

Abstract methods were annotated with @Nonnull, as well as
TypeQualifierValue<?>s and Column Types.

Fix checkstyle issues.

Make findSuperMethods private on UnknownNullnessDetector
The method was public but should be private.

Test Plan: No test plan.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Subscribers: itirabasso

Differential Revision: http://ph.monits.com/D14950